Switched LCD and emulated Tom's code
Lots of different changes we basically took Tom's code from his 8bit
micro  bit micro
double check timing requirements
we saw a "heartbeat" of an x and a m display and switch back and forth
on the display.
We should've seen an o instead of an m. Check that bits are correct, and
that timing is correct.